Move UniformChar, UniformDuration to new submodule
